+++ projects/ui-lib/footers/columns-footer/columns-footer.interface.ts
@@ -1,0 +1,6 @@
+import { ColumnsFooterCopyRightI, ColumnsFooterItemI } from "./columns-footer.component";
+
+export interface ColumnsFooterI {
+    columns?: ColumnsFooterItemI[];
+    copyright?: ColumnsFooterCopyRightI;
+}