+++ projects/ui-lib/footers/columns-footer/index.ts
@@ -1,0 +1,2 @@
+export * from './columns-footer.component';
+export * from './columns-footer.interface';