+++ projects/ui-lib/mapper/cloudinary.token.ts
@@ -1,0 +1,3 @@
+import { InjectionToken } from '@angular/core';
+
+export const CLOUDINARY_CLOUD_NAME = new InjectionToken<string>('CLOUDINARY_CLOUD_NAME');