+++ projects/ui-lib/shared/ui-image/ui-image.interface.ts
@@ -1,0 +1,5 @@
+export interface UiImageI {
+    publicId: string;
+    cloudName?: string;
+    transformations?: any;
+}